requirements.txt
environ
djangorestframework
django-filter
asgiref==3.3.1
Brotli==1.0.9
Django==3.1.3
django-appconf==1.0.4
django-constance==2.8.0
django-cors-headers==3.7.0
django-image-cropping==1.5.0
django-picklefield==3.0.1
django-resized==0.3.11
fonttools==4.16.1
lxml @ file:///C:/Users/Michael%20Postnikov/Downloads/lxml-4.5.2-cp38-cp38-win32.whl
Pillow==8.1.0
pytz==2020.4
sqlparse==0.4.1
zopfli==0.1.7
